+++ services/one-app/src/model/Auth.ts
@@ -1,0 +1,10 @@
+export enum SocialSignInType {
+  KAKAO = 'KAKAO',
+  APPLE = 'APPLE',
+  GOOGLE = 'GOOGLE',
+}
+
+export enum CookieKey {
+  ACCESS_TOKEN = 'AHAHCHUL_ACCESS_TOKEN',
+  REFRESH_TOKEN = 'AHAHCHUL_REFRESH_TOKEN',
+}
